nlldo.c: fix NDM_LOG_ERROR warnings
nlldo.c: In function 'nlldo_handle_packet':
nlldo.c:224:18: warning: format '%u' expects argument of type 'unsigned int',
but argument 3 has type 'size_t {aka long unsigned int}' [-Wformat=]
    NDM_LOG_ERROR("unable to receive LLDP packet: %u", bytes_read);
                  ^
../libndm/include/ndm/log.h:60:54: note: in definition of macro 'NDM_LOG_ERROR'
 #define NDM_LOG_ERROR(fmt, args...)  ndm_log(LERROR, fmt, ##args)
                                                      ^~~
nlldo.c:263:19: warning: format '%d' expects argument of type 'int',
but argument 3 has type 'size_t {aka long unsigned int}' [-Wformat=]
     NDM_LOG_ERROR("packet len: %d, offset: %d, TLV len:%d",
                   ^
../libndm/include/ndm/log.h:60:54: note: in definition of macro 'NDM_LOG_ERROR'
 #define NDM_LOG_ERROR(fmt, args...)  ndm_log(LERROR, fmt, ##args)
                                                      ^~~
nlldo.c:263:19: warning: format '%d' expects argument of type 'int',
but argument 4 has type 'long int' [-Wformat=]
     NDM_LOG_ERROR("packet len: %d, offset: %d, TLV len:%d",
                   ^
../libndm/include/ndm/log.h:60:54: note: in definition of macro 'NDM_LOG_ERROR'
 #define NDM_LOG_ERROR(fmt, args...)  ndm_log(LERROR, fmt, ##args)
                                                      ^~~

+++ nlldo.c
@@ -221,7 +221,7 @@ static void nlldo_handle_packet()
 		bytes_read == 0) {
 
 		if (debug) {
-			NDM_LOG_ERROR("unable to receive LLDP packet: %u", bytes_read);
+			NDM_LOG_ERROR("unable to receive LLDP packet: %zu", bytes_read);
 		}
 
 		return;
@@ -260,8 +260,8 @@ static void nlldo_handle_packet()
 
 		if (OFFSET(packet, p) + TLV_HDR_LEN + TLV_LEN(tlv->hdr) > bytes_read) {
 			if (debug) {
-				NDM_LOG_ERROR("packet len: %d, offset: %d, TLV len:%d",
-					bytes_read, OFFSET(packet, p), TLV_LEN(tlv->hdr));
+				NDM_LOG_ERROR("packet len: %zu, offset: %zu, TLV len:%zu",
+					bytes_read, (size_t) OFFSET(packet, p), (size_t) TLV_LEN(tlv->hdr));
 			}
 
 			break;
